patches rerender problem for demo
