+++ box-ui/src/components/tests/JitsiComponent.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import JitsiFrame from '../ts/JitsiFrame';
 
 test('renders learn react link', () => {
-    render(<JitsiFrame information={{ domain: '', roomName: '' }} options={{}} />);
+    render(<JitsiFrame information={{ domain: '', roomName: '' }} showYellowbg={false} options={{}} />);
     const linkElement = screen.getByText(/learn react/i);
     expect(linkElement).toBeInTheDocument();
 });
