+++ box-ui/src/components/ts/Box/BoxMeeting.tsx
@@ -40,58 +40,15 @@ const BoxMeeting: FunctionComponent = () => {
         }),
         [],
     );
-    const [raised, setRaised] = useState(false);
-    const [counter, setCounter] = useState(0);
-    const TimeOutRef = useRef<NodeJS.Timeout>();
-
-    function lowerHand() {
-        setRaised(false);
-        setCounter(0);
-    }
-
-    const updateCounter = useCallback((timeRaised: number) => {
-        if (timeRaised > 0) {
-            setCounter((counter) => counter + 1);
-        }
-        if (timeRaised == 0) {
-            setCounter((counter) => Math.max(0, counter - 1));
-        }
-    }, []);
-
-    function switchHand() {
-        let stop = false;
-        if (counter > 0) {
-            stop = true;
-        }
-        if (stop) {
-            setRaised(true);
-        } else {
-            setRaised(false);
-        }
-    }
-    useEffect(() => {
-        const id = setTimeout(lowerHand, 10000);
-        if (TimeOutRef.current) {
-            clearTimeout(TimeOutRef.current);
-        }
-        TimeOutRef.current = id;
-        switchHand();
-    }, [counter]);
 
     return (
         <div className='BoxMeeting'>
             <PopupComponent information={information} setInformation={setInformation} />
             <div className='CreateMeetingContainer'>
                 <div className='JitsiComponent'>
-                    <div
-                        className='overlay'
-                        style={{
-                            backgroundColor: 'yellow',
-                            opacity: raised ? '0.5' : '0',
-                        }}
-                    />
                     <JitsiFrame
                         information={information}
+                        showYellowbg={true}
                         options={meetingOptions}
                         configure={(api) => {
                             api.addListener('videoConferenceLeft', () => {
@@ -103,10 +60,6 @@ const BoxMeeting: FunctionComponent = () => {
                                         domain: information.domain,
                                     },
                                 });
-                            });
-                            api.addListener('raiseHandUpdated', (res) => {
-                                const timeRaised = res.handRaised;
-                                updateCounter(timeRaised);
                             });
                         }}
                         onError={() => {
