+++ box-ui/src/utils/Props/index.d.ts
@@ -37,6 +37,7 @@ interface InformationProps {
 
 export type JitsiFrameProps = {
     information: Information;
+    showYellowbg : boolean;
     options: ConstructorParameters<typeof JitsiMeetExternalAPI>[1];
     configure?: (api: JitsiMeetExternalAPI) => void;
     onError?: () => void;
